move read_frame_header_and_toc
